ares-x-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.4.0
redis==5.0.4
langchain==0.2.1
langchain-openai==0.1.7
langgraph==0.0.55
python-dotenv==1.0.1
httpx==0.27.0
pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
alembic==1.13.1
prometheus-client==0.20.0
